refactor: Borrar ActivityStudentUI. #243
