--- src/student/context/activityStudentContext/activityStudentContextUI/ActivityStudentContextUI.ts
@@ -1,6 +1,0 @@
-import { createContext } from "react";
-import type { ActivityStudentContextUIType } from "./ActivityStudentContextUI.type";
-
-export const ActivityStudentContextUI = createContext<
-  ActivityStudentContextUIType | undefined
->(undefined);
